planne-sdk/pyproject.toml
[project]
name = "planne-sdk"
version = "0.1.0"
description = "SDK for Planne's projects. Contains data models and use cases for interacting with Planne's DBs and systems, while following all the business logic. To be used by any other project in the eco-system. Follows semantic versioning."
readme = "README.md"
authors = [
    { name = "vinivosh", email = "36175157+vinivosh@users.noreply.github.com" }
]
requires-python = ">=3.11.14"
dependencies = [
    "alembic[tz]==1.18.4",
    "psycopg-binary==3.3.2",
    "pydantic[email]==2.12.5",
    "python-dotenv==1.2.1",
    "sqlmodel==0.0.33",
    "structlog==25.5.0",
]

[build-system]
requires = ["uv_build==0.10.1,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "commitlint==1.13.3",
    "pre-commit==4.5.1",
    "pytest==9.0.2",
    "rich==14.3.2",
    "ruff==0.15.1",
]
migrations = [
    "alembic[tz]==1.18.4",
    "psycopg==3.3.2",
]


[tool.ruff]
line-length = 79


[tool.ruff.lint]
extend-select = [
    "D",     # pydocstyle
    "E501",  # line-too-long
    "N",     # PEP8 naming
    "UP",    # pyupgrade
]
extend-ignore = [
    "E501",  # line-too-long
    "F403",  # 'from module import *' used; unable to detect undefined names
]


[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.alembic]

# path to migration scripts.
# this is typically a path given in POSIX (e.g. forward slashes)
# format, relative to the token %(here)s which refers to the location of this
# ini file
script_location = "%(here)s/alembic"

# template used to generate migration file names; The default value is %%(rev)s_%%(slug)s
# Uncomment the line below if you want the files to be prepended with date and time
# see https://alembic.sqlalchemy.org/en/latest/tutorial.html#editing-the-ini-file
# for all available tokens
file_template = "%%(year)d_%%(month).2d_%%(day).2d_%%(hour).2d%%(minute).2d-%%(rev)s_%%(slug)s"
# Or organize into date-based subdirectories (requires recursive_version_locations = true)
# file_template = "%%(year)d/%%(month).2d/%%(day).2d_%%(hour).2d%%(minute).2d_%%(second).2d_%%(rev)s_%%(slug)s"

# additional paths to be prepended to sys.path. defaults to the current working directory.
prepend_sys_path = [
    "."
]

# timezone to use when rendering the date within the migration file
# as well as the filename.
# If specified, requires the tzdata library which can be installed by adding
# `alembic[tz]` to the pip requirements.
# string value is passed to ZoneInfo()
# leave blank for localtime
timezone = "UTC"

# max length of characters to apply to the "slug" field
# truncate_slug_length = 40

# set to 'true' to run the environment during
# the 'revision' command, regardless of autogenerate
# revision_environment = false

# set to 'true' to allow .pyc and .pyo files without
# a source .py file to be detected as revisions in the
# versions/ directory
# sourceless = false

# version location specification; This defaults
# to <script_location>/versions.  When using multiple version
# directories, initial revisions must be specified with --version-path.
# version_locations = [
#    "%(here)s/alembic/versions",
#    "%(here)s/foo/bar"
# ]


# set to 'true' to search source files recursively
# in each "version_locations" directory
# new in Alembic version 1.10
# recursive_version_locations = false

# the output encoding used when revision files
# are written from script.py.mako
# output_encoding = "utf-8"

# This section defines scripts or Python functions that are run
# on newly generated revision scripts.  See the documentation for further
# detail and examples
# [[tool.alembic.post_write_hooks]]
# format using "black" - use the console_scripts runner,
# against the "black" entrypoint
# name = "black"
# type = "console_scripts"
# entrypoint = "black"
# options = "-l 79 REVISION_SCRIPT_FILENAME"
#
[[tool.alembic.post_write_hooks]]
# Lint with ruff
name = "ruff"
type = "module"
module = "ruff"
options = "format REVISION_SCRIPT_FILENAME"

[[tool.alembic.post_write_hooks]]
# Attempt to fix with ruff
name = "ruff"
type = "module"
module = "ruff"
options = "check --fix REVISION_SCRIPT_FILENAME"

# [[tool.alembic.post_write_hooks]]
# Alternatively, use the exec runner to execute a binary found on your PATH
# name = "ruff"
# type = "exec"
# executable = "ruff"
# options = "check --fix REVISION_SCRIPT_FILENAME"

